+++ OmegaNum.js
@@ -78,7 +78,11 @@
    *  isNaN
    *  isNegative                isneg
    *  isPositive                ispos
+   *  iteratedexp
+   *  iteratedlog
    *  lambertw
+   *  layeradd
+   *  layeradd10
    *  lessThan                  lt
    *  lessThanOrEqualTo         lte
    *  logarithm                 logBase
@@ -655,9 +659,12 @@
   };
   //end break_eternity.js excerpt
   //Uses linear approximations for real height
-  P.tetrate=P.tetr=function (other){
+  P.tetrate=P.tetr=function (other,payload){
+    if (payload===undefined) payload=OmegaNum.ONE;
     var t=this.clone();
     other=new OmegaNum(other);
+    payload=new OmegaNum(payload);
+    if (payload.neq(OmegaNum.ONE)) other=other.add(payload.slog(t));
     if (OmegaNum.debug>=OmegaNum.NORMAL) console.log(t+"^^"+other);
     var negln;
     if (other.isInfinite()&&other.sign>0){
@@ -724,9 +731,46 @@
     r.standardize();
     return r;
   };
-  Q.tetrate=Q.tetr=function (x,y){
-    return new OmegaNum(x).tetr(y);
+  Q.tetrate=Q.tetr=function (x,y,payload){
+    return new OmegaNum(x).tetr(y,payload);
   };
+  //Implementation of functions from break_eternity.js
+  P.iteratedexp=function (other,payload){
+    return this.tetr(other,payload);
+  };
+  Q.iteratedexp=function (x,y,payload){
+    return new OmegaNum(x).iteratedexp(other,payload);
+  }
+  //This implementation is highly inaccurate and slow, and probably be given custom code
+  P.iteratedlog=function (base,other){
+    if (base===undefined) base=10;
+    if (other===undefined) other=OmegaNum.ONE.clone();
+    var t=this.clone();
+    base=new OmegaNum(base);
+    other=new OmegaNum(other);
+    return base.tetr(t.slog(base).sub(other));
+  }
+  Q.iteratedlog=function (x,y,z){
+    return new OmegaNum(x).iteratedlog(y,z);
+  }
+  P.layeradd=function (other,base){
+    if (base===undefined) base=10;
+    if (other===undefined) other=OmegaNum.ONE.clone();
+    var t=this.clone();
+    base=new OmegaNum(base);
+    other=new OmegaNum(other);
+    return base.tetr(t.slog(base).add(other));
+  }
+  Q.layeradd=function (x,y,z){
+    return new OmegaNum(x).layeradd(y,z);
+  }
+  P.layeradd10=function (other){
+    return this.layeradd(other);
+  }
+  Q.layeradd10=function (x,y){
+    return new OmegaNum(x).layeradd10(y);
+  }
+  //End implementation from break_eternity.js
   //All of these are from Patashu's break_eternity.js
   //The super square-root function - what number, tetrated to height 2, equals this?
   //Other sroots are possible to calculate probably through guess and check methods, this one is easy though.